Correct range check scanning for an XROM entry
While looking for an XROM entry, the check whether a given entry
exists was wrong. It would "find" the first entry that is just outside
the FAT table, causing the FAT trailer to incorrectly be used as an
entry.

Now it will properly see that it is out of range and keep looking for
the XROM entry in the following pages.