Correct problem in GTRMAD
The idea of GTRMAD is to scan and locate an 'XROM i,j' function.

In case there are multiple XROMs matching 'i', the first one that can
provide a function 'j' is to be used. At least that is the idea, but
for 40 years it has never worked. The problem was that when the first
XROM 'i' was found with too few functions, it did not properly restore
registers for a continued search, trashing the 'i' value it compared
with. Thus, as a result not being able to locate 'XROM i,j' even if
was present further down the line.