Replicate NEWT image with timing annotations
Provide timing annotations for NEWT. This does not follow the existing
41CL image exactly. I took the liberty to tune it a bit after the best
of my ability, reducing the amount of slow-downs and also make one
that was slightly off to have better cover (DRSY26).

Provide additional 41CL routines found in the extended functions and
TIME module, including the 'YRES' instruction.

The cold start MMU disable is done in a different way here. The
existing 41CL image makes use of the cold start vector in the Time
module, but uses a relay jump in the ID area which trashes it. I opted
for having the module ID intact and the cold start routine is modified
to reset the NEWT MMU before giving control to poll vectors.

The module ID of Extended Functions is properly bumped here.