Bump mainframe to revision IFG
The idea is that the last version was the one following HFF, but it
was actually named NFL (probably by some football fan).

To avoid this version history gap, we rewind the history here and
start off from the next version after HFF. Later we can just skip the
N and L versions so we get a reasonable long series of meaningful
version numbers. This means that the next version is expected to be
IFG (unless some correction is done to that mid ROM later).