Initial support for more extended memory
Make use of NET memory in the 400-FFF range to support additional
extended memory pages. In this case we make use of the full page
without any void gaps. A #define makes it easy to configure it for any
number of additional pages.

It is so far very lightly tested. If using up to page 14 (which leaves
page 15 free for other purposes), 3405 registers of extended memory
are available. This means about 23K of extended memory.