Speed up cold starts
As pointed out long ago by DRC, there is no need to clear page 2 and
3. At that point it might not have been clear what these pages would
be used for, but we know it is for extended memory. It is only
important to clear the first module which resides in part of page 0.
Hence, if suffice to clear up to 0x1ff.
The alternative would be to clear even more page on the HP-41CL, but
as it seems not needed, be consistent and allow for faster cold
starts.