Update EMDIR
Support showing file sizes >999. Also display more file types.
The EMDIR2 label change meaning here. This should not be a problem as
it is not really a proper entry point, it is just a continuation in
the middle of the EMDIR code and now it had to be cut a bit
differently. As it not longer have the same meaning, it is also moved
to have a single unused fill block.