Disable Time module speed annotations on DADD=C
Running the DADD=C instruction on the bus to disable the Timer chip
turns out to be unnecessary as it does not matter that the Timer
thinks it is active. This is because when RAM is active, the access is
internal to the NEWT and not shown on the bus. Hence, the potentially
still active Timer chip has no idea that there are RAM read/write
access going on, as it cannot see it.