Fix CAT 2 speed in turbo mode
CAT 2 ran too fast when using turbo mode. It was totally impossible to
see where it was when a key was held down to speed it up further.
The problem is related to that the delay loop is in two parts, the
loop itself and a CAT_STOP subroutine which was not speed annotated
before.