Correct indirect stack prompts for XROMs
An XROM can request numeric input (partial key sequence). In this case
the postfix byte should not be merged with the instruction (as an XROM
is two bytes), and instead be delivered in B.X.

This is done correctly in most cases, but for stack indirect inputs,
it would incorrectly merge the instruction, overwriting the second
byte of the XROM function code, either causing a NONEXISTENT or
execution of a different instruction, which potentially can be quite
dangerous.